ci: fix atoi.v's max_u64

+++ vlib/strconv/atoi.v
@@ -9,7 +9,7 @@ const (
 // int_size = 32 << (~u32(0) >> 63)
 // max_u64 = u64(u64(1 << 63) - 1)
 	int_size = 32
-	max_u64 = 18446744073709551615 // as u64 // use this until we add support
+	max_u64 = u64(18446744073709551615) // as u64 // use this until we add support
 )
 
 pub fn byte_to_lower(c byte) byte {
